The slippery pthread support for systems prior to 502102 has been dropped and the lang/ruby16_r and lang/ruby18_r ports have been removed, since no one seems to appreciate the partially working solution.
Good news is that the pthread support of lang/ruby18 is now enabled by
default for newer systems, which means the ruby interpreter is linked
with libpthread.  This will allow threaded extension libraries to run
and work properly on those systems.

The --march=cputype flag is disabled because it gets ruby to
malfunction and fail to build.  I don't know if the problem is in
libpthread or in gcc.

(It really makes me wonder if they had actually tested before asking
me to do this somewhat risky change ;-)