The PEAR::Net_Ident implements Identification Protocol according to RFC 1413.
The Identification Protocol (a.k.a., "ident", a.k.a., "the Ident
Protocol") provides a means to determine the identity of a user
of a particular TCP connection. Given a TCP port number pair, it
returns a character string which identifies the owner of that
connection on the server's system.

PR:		ports/78532
Submitted by:	Antonio Carlos Venancio Junior <antonio@php.net>