Add wavpack.
Wavepack is a completely open audio compression format providing lossless,
high-quality lossy, and a unique hybrid compression mode.
In the default lossless mode no original information is lost, so there is no
change of degradation. Compression in this mode is between 30% an 70%.

For more info see the website http://www.wavpack.com