Add feappv 2.0, finite Element Analysis Program "personal version".
This is a FEA program used in a classic FEM book. A complete (commercial)
version is available here:

	<http://www.ce.berkeley.edu/~rlt/feap/>

The "personal version" is very limited, but it keeps the same format as
the complete (commercial) version and cad/netgen can produce files for it.

PR:		ports/95210
Submitted by:	Pedro F. Giffuni <giffunip (at) asme.org>