While trying to upgrade devel/tcllib several test failures were detected in the RIPEMD128 implementation.
Turns out, the problem was fixed a few months ago by the TRF author in CVS.
This commit adds a patch and bumps portrevision.

Although the same problem existed in the TRF's RIPEMD160 implementation,
it did not affect this port, because it uses the OpenSSL's (or -lmd's)
implementations of all digests -- except those, which are considered
obsolete and no longer shipped. Like rmd128.

Thanks to:	Denis Shaposhnikov